requirements.txt
aerich==0.5.3
astroid==2.5.6
asyncpg==0.23.0
autopep8==1.5.7
bcrypt==3.2.0
cffi==1.14.5
cryptography==3.4.7
ddlparse==1.9.0
Deprecated==1.2.12
dictdiffer==0.8.1
dnspython==2.1.0
email-validator==1.1.2
fastapi==0.65.1
fastapi-crudrouter==0.7.1
fastapi-pagination==0.7.3
greenlet==1.1.0
h11==0.12.0
idna==3.2
iso8601==0.1.14
isort==5.8.0
lazy-object-proxy==1.6.0
makefun==1.11.3
mccabe==0.6.1
orjson==3.5.2
passlib==1.7.4
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
pylint==2.8.2
pyparsing==2.4.7
python-dotenv==0.17.1
python-jose==3.2.0
python-multipart==0.0.5
pytz==2021.1
rsa==4.7.2
six==1.16.0
starlette==0.14.2
toml==0.10.2
tortoise-orm==0.17.3
typing-extensions==3.10.0.0
uvicorn==0.13.4
wrapt==1.12.1
